+++ objectified-ui/public/WHATS_NEW.md
@@ -119,6 +119,7 @@ We continue to improve the platform based on your feedback with improvements and
     - An overall of all detected conflicts
     - In the impact analysis portion, show what will change if resolved
     - Adds Duplicate Schema Detection
+    - Adds detection of property conflicts: incompatible property definitions
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions
   - Improved import support for paths and securitySchemes